M1-flow-azure/requirements.txt
azure-functions
azure-data-tables
azure-storage-queue
azure-storage-blob
azure-identity
azure-keyvault-secrets
numpy
pandas
scikit-learn
lightgbm
onnxruntime
boto3
python-json-logger
requests
pydantic
pydantic-settings
onnxmltools
simpy
flask
locust

# AI/OpenAI Integration
openai>=1.12.0
tiktoken
tenacity
python-dotenv
